+++ app/frontend/webserver/handlers/__tests__/restore-post.test.js
@@ -1,0 +1,194 @@
+/* eslint-disable global-require */
+describe('webserver/handlers/restore-post', () => {
+  let loggerMock;
+
+  let shellMock;
+
+  let rsSend;
+  let archiveMv;
+
+  const appMock = {
+    route: null
+  };
+
+  let postReq;
+
+  let mockVerifyAuthenticated;
+
+  beforeEach(async () => {
+    jest.clearAllMocks().resetModules();
+
+    const shell = require('shelljs');
+    jest.mock('shelljs');
+
+    shellMock = shell;
+    shellMock.exec = jest.fn().mockImplementation((_cmd, fn) => fn());
+
+    rsSend = jest.fn().mockResolvedValue(true);
+    appMock.route = jest.fn(() => ({
+      post: jest.fn().mockImplementation(func => {
+        func(postReq, { send: rsSend });
+      })
+    }));
+  });
+
+  describe('when verification failed', () => {
+    beforeEach(async () => {
+      const { logger } = require('../../../../helpers');
+
+      loggerMock = logger;
+
+      mockVerifyAuthenticated = jest.fn().mockResolvedValue(false);
+
+      jest.mock('../../../../cronjob/trailingTradeHelper/common', () => ({
+        verifyAuthenticated: mockVerifyAuthenticated
+      }));
+
+      postReq = {
+        header: () => 'some token'
+      };
+      const { handleRestorePost } = require('../restore-post');
+
+      await handleRestorePost(loggerMock, appMock);
+    });
+
+    it('triggers verifyAuthenticated', () => {
+      expect(mockVerifyAuthenticated).toHaveBeenCalledWith(
+        loggerMock,
+        'some token'
+      );
+    });
+
+    it('return unauthorised', () => {
+      expect(rsSend).toHaveBeenCalledWith({
+        success: false,
+        status: 403,
+        message: 'Please authenticate first.',
+        data: {}
+      });
+    });
+  });
+
+  describe('when verification success', () => {
+    beforeEach(() => {
+      archiveMv = jest.fn();
+    });
+
+    describe(`backup failed`, () => {
+      beforeEach(async () => {
+        const { logger } = require('../../../../helpers');
+
+        loggerMock = logger;
+
+        mockVerifyAuthenticated = jest.fn().mockResolvedValue(true);
+
+        jest.mock('../../../../cronjob/trailingTradeHelper/common', () => ({
+          verifyAuthenticated: mockVerifyAuthenticated
+        }));
+
+        postReq = {
+          header: () => 'some token',
+          files: {
+            archive: {
+              name: 'my-backup.archive',
+              mv: archiveMv
+            }
+          }
+        };
+
+        shellMock.exec = jest
+          .fn()
+          .mockImplementation((_cmd, fn) => fn(1, '', 'something happened'));
+
+        const { handleRestorePost } = require('../restore-post');
+
+        await handleRestorePost(loggerMock, appMock);
+      });
+
+      it('triggers verifyAuthenticated', () => {
+        expect(mockVerifyAuthenticated).toHaveBeenCalledWith(
+          loggerMock,
+          'some token'
+        );
+      });
+
+      it('moves to tmp folder', () => {
+        expect(archiveMv).toHaveBeenCalled();
+      });
+
+      it('return failed', () => {
+        expect(rsSend).toHaveBeenCalledWith({
+          success: false,
+          status: 500,
+          message: 'Restore failed',
+          data: {
+            code: 1,
+            stderr: 'something happened',
+            stdout: ''
+          }
+        });
+      });
+    });
+
+    describe(`backup succeeed`, () => {
+      beforeEach(async () => {
+        const { logger } = require('../../../../helpers');
+
+        loggerMock = logger;
+
+        mockVerifyAuthenticated = jest.fn().mockResolvedValue(true);
+
+        jest.mock('../../../../cronjob/trailingTradeHelper/common', () => ({
+          verifyAuthenticated: mockVerifyAuthenticated
+        }));
+
+        postReq = {
+          header: () => 'some token',
+          files: {
+            archive: {
+              name: 'my-backup.archive',
+              mv: archiveMv
+            }
+          }
+        };
+
+        shellMock.exec = jest
+          .fn()
+          .mockImplementation((_cmd, fn) => fn(0, 'all good', ''));
+
+        const { handleRestorePost } = require('../restore-post');
+
+        await handleRestorePost(loggerMock, appMock);
+      });
+
+      it('triggers verifyAuthenticated', () => {
+        expect(mockVerifyAuthenticated).toHaveBeenCalledWith(
+          loggerMock,
+          'some token'
+        );
+      });
+
+      it('triggers shell.exec', () => {
+        expect(shellMock.exec).toHaveBeenCalledWith(
+          expect.stringContaining(
+            `${process.cwd()}/scripts/restore.sh binance-mongo 27017`
+          ),
+          expect.any(Function)
+        );
+      });
+
+      it('return success', () => {
+        expect(rsSend).toHaveBeenCalledWith({
+          success: true,
+          status: 200,
+          message: 'Restore success',
+          data: {
+            code: 0,
+            stderr: '',
+            stdout: 'all good'
+          }
+        });
+      });
+    });
+  });
+});